FAST_API_/requirements.txt
fastapi
uvicorn
python-dotenv
sentence-transformers
numpy
requests
pydantic